MDEXP-487 Fix unit test

+++ src/main/java/org/folio/service/file/cleanup/StorageCleanupServiceImpl.java
@@ -55,8 +55,10 @@ public class StorageCleanupServiceImpl implements StorageCleanupService {
     Promise<Boolean> promise = Promise.promise();
     return fileStorage.deleteFileAndParentDirectory(fileDefinition)
       .compose(isFileDeleted -> {
-        fileDefinitionDao.deleteById(fileDefinition.getId(), tenantId);
-        return Future.succeededFuture(true);
+        if (Boolean.TRUE.equals(isFileDeleted)) {
+          return fileDefinitionDao.deleteById(fileDefinition.getId(), tenantId);
+        }
+        return Future.succeededFuture(false);
       })
       .compose(isFileDefinitionDeleted -> {
         if (Boolean.FALSE.equals(isFileDefinitionDeleted)) {
